rename: 모델 하위의 파일 패키지 이동
