+++ src/main/java/model/player/Player.java
@@ -1,6 +1,8 @@
-package model;
+package model.player;
 
 import java.util.List;
+import model.card.Card;
+import model.card.Cards;
 
 public class Player {
